Refactor: Flatten module hierarchy and promote bard highlighter
