+++ src/ast.rs
@@ -2,6 +2,37 @@
 //!
 //! These nodes capture the structure of a GLOSSA program,
 //! preserving both original Greek text and normalized forms.
+//!
+//! # Structure
+//!
+//! * [`Program`]: The root node, containing a list of [`Statement`]s.
+//! * [`Statement`]: A top-level instruction, ending with a period (`.`) or query mark (`?` / `;`).
+//!   * A statement consists of one or more [`Clause`]s.
+//! * [`Clause`]: A comma-separated part of a statement.
+//!   * Example: `ὁ ἄνθρωπος, τὸν λόγον λέγει.` (Two clauses: "The man", "says the word").
+//! * [`Expr`]: An expression (word, literal, operation).
+//!   * [`Expr::Word`]: A raw Greek word with its original and normalized forms.
+//!
+//! # Design Philosophy
+//!
+//! Unlike traditional ASTs that might discard surface-level details, the GLOSSA AST
+//! preserves the *original* Greek text in [`Word`] nodes. This is crucial for:
+//!
+//! 1. **Error Reporting**: Using the original polytonic Greek in error messages.
+//! 2. **Morphological Analysis**: The semantic phase needs the original form to
+//!    distinguish subtle variations if needed.
+//!
+//! # Example
+//!
+//! A simple program like `«χαῖρε» λέγε.` produces:
+//!
+//! ```text
+//! Program
+//! └── Statement::Regular
+//!     └── Clause
+//!         ├── Expr::StringLiteral("χαῖρε")
+//!         └── Expr::Word("λέγε")
+//! ```
 
 use smol_str::SmolStr;
