+++ src/report.rs
@@ -7,9 +7,7 @@ use comfy_table::{Cell, Color, Table, presets};
 use crossterm::style::Stylize;
 use std::fmt::Display;
 
-use crate::semantic::{
-    AnalyzedExpr, AnalyzedExprKind, AnalyzedProgram, AnalyzedStatement,
-};
+use crate::semantic::{AnalyzedExpr, AnalyzedExprKind, AnalyzedProgram, AnalyzedStatement};
 
 /// Statistics for an analyzed program
 #[derive(Debug, Default, Clone)]
@@ -37,12 +35,12 @@ pub struct ProgramStats {
 impl ProgramStats {
     /// Analyze a program and collect statistics
     pub fn new(program: &AnalyzedProgram) -> Self {
-        let mut stats = ProgramStats::default();
-
-        // Count top-level definitions from scope
-        stats.function_count = program.scope.functions().count();
-        stats.type_count = program.scope.types().count();
-        stats.trait_count = program.scope.traits().count();
+        let mut stats = ProgramStats {
+            function_count: program.scope.functions().count(),
+            type_count: program.scope.types().count(),
+            trait_count: program.scope.traits().count(),
+            ..ProgramStats::default()
+        };
 
         // Traverse statements to count structural elements
         for stmt in &program.statements {
@@ -71,7 +69,11 @@ impl ProgramStats {
                     self.visit_expr(expr);
                 }
             }
-            AnalyzedStatement::If { condition, then_body, else_body } => {
+            AnalyzedStatement::If {
+                condition,
+                then_body,
+                else_body,
+            } => {
                 self.conditional_count += 1;
                 self.visit_expr(condition);
                 for s in then_body {
@@ -157,8 +159,11 @@ impl ProgramStats {
                     self.visit_expr(e);
                 }
             }
-            AnalyzedExprKind::Some(e) | AnalyzedExprKind::Ok(e) | AnalyzedExprKind::Err(e)
-            | AnalyzedExprKind::Unwrap(e) | AnalyzedExprKind::Try(e) => {
+            AnalyzedExprKind::Some(e)
+            | AnalyzedExprKind::Ok(e)
+            | AnalyzedExprKind::Err(e)
+            | AnalyzedExprKind::Unwrap(e)
+            | AnalyzedExprKind::Try(e) => {
                 self.visit_expr(e);
             }
             AnalyzedExprKind::IndexAccess { array, index } => {
@@ -212,11 +217,12 @@ impl<'a> GlossaReport<'a> {
 impl Display for GlossaReport<'_> {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         let mut table = Table::new();
-        table.load_preset(presets::UTF8_FULL)
-             .set_header(vec![
-                 Cell::new("Μετρική (Metric)").add_attribute(comfy_table::Attribute::Bold).fg(Color::Cyan),
-                 Cell::new("Τιμή (Value)").add_attribute(comfy_table::Attribute::Bold),
-             ]);
+        table.load_preset(presets::UTF8_FULL).set_header(vec![
+            Cell::new("Μετρική (Metric)")
+                .add_attribute(comfy_table::Attribute::Bold)
+                .fg(Color::Cyan),
+            Cell::new("Τιμή (Value)").add_attribute(comfy_table::Attribute::Bold),
+        ]);
 
         table.add_row(vec![
             Cell::new("Ἀρχεῖον (File)"),
@@ -236,14 +242,14 @@ impl Display for GlossaReport<'_> {
         ]);
 
         if self.stats.function_count > 0 {
-             table.add_row(vec![
+            table.add_row(vec![
                 Cell::new("Συναρτήσεις (Functions)"),
                 Cell::new(self.stats.function_count),
             ]);
         }
 
         if self.stats.type_count > 0 {
-             table.add_row(vec![
+            table.add_row(vec![
                 Cell::new("Τύποι (Types)"),
                 Cell::new(self.stats.type_count),
             ]);
@@ -251,19 +257,23 @@ impl Display for GlossaReport<'_> {
 
         if self.stats.loop_count > 0 {
             table.add_row(vec![
-               Cell::new("Βρόχοι (Loops)"),
-               Cell::new(self.stats.loop_count),
-           ]);
-       }
+                Cell::new("Βρόχοι (Loops)"),
+                Cell::new(self.stats.loop_count),
+            ]);
+        }
 
-       if self.stats.max_depth > 0 {
-        table.add_row(vec![
-           Cell::new("Βάθος (Max Depth)"),
-           Cell::new(self.stats.max_depth),
-       ]);
-   }
+        if self.stats.max_depth > 0 {
+            table.add_row(vec![
+                Cell::new("Βάθος (Max Depth)"),
+                Cell::new(self.stats.max_depth),
+            ]);
+        }
 
-        writeln!(f, "\n{}", "ΑΝΑΦΟΡΑ ΓΛΩΣΣΗΣ (LANGUAGE REPORT)".bold().underlined())?;
+        writeln!(
+            f,
+            "\n{}",
+            "ΑΝΑΦΟΡΑ ΓΛΩΣΣΗΣ (LANGUAGE REPORT)".bold().underlined()
+        )?;
         writeln!(f, "{}", table)?;
 
         // If there are top-level functions, list them
@@ -271,16 +281,25 @@ impl Display for GlossaReport<'_> {
         if !functions.is_empty() {
             writeln!(f, "\n{}", "ΣΥΝΑΡΤΗΣΕΙΣ (FUNCTIONS)".bold())?;
             let mut func_table = Table::new();
-            func_table.load_preset(presets::UTF8_HORIZONTAL_ONLY)
-                .set_header(vec!["Ὄνομα (Name)", "Παράμετροι (Params)", "Επιστροφή (Returns)"]);
+            func_table
+                .load_preset(presets::UTF8_HORIZONTAL_ONLY)
+                .set_header(vec![
+                    "Ὄνομα (Name)",
+                    "Παράμετροι (Params)",
+                    "Επιστροφή (Returns)",
+                ]);
 
             for func in functions {
-                let params = func.param_types.iter()
+                let params = func
+                    .param_types
+                    .iter()
                     .map(|t| t.to_string())
                     .collect::<Vec<_>>()
                     .join(", ");
 
-                let ret = func.return_type.as_ref()
+                let ret = func
+                    .return_type
+                    .as_ref()
                     .map(|t| t.to_string())
                     .unwrap_or_else(|| "Οὐδέν".to_string());
